Update README.md
Dino Dash is a fun and challenging game where players control a dinosaur trying to survive an asteroid storm. The objective is simple but thrilling: avoid the asteroids coming one after another to keep the dinosaur safe! With increasing difficulty, the game tests players' reflexes and timing skills.

Features : 
- Play as a dinosaur navigating through waves of incoming asteroids.
- Increasing difficulty with each level to keep players engaged.
- Simple controls and addictive gameplay.

This project is perfect for anyone interested in game development, collision detection, and UI design. Dive in and help the dinosaur survive the ultimate space adventure!